Had to save the envir ontology in RDF as Jena does not like OWL.
Ontologies are loaded on startup. To minimise loading time, in time for pilot we should only load which onts we actually need. Direct imported onts will be loaded regardless, but this will reduce loading time, and should reduce memory footprint.

Do a GC after loading onts to free a few megs of memory. Consider this if we get problems.